fix(bevy): cast sim_speed to f32 for GameSpeed.multiplier (wave1↔green drift)

+++ clients/bevy-ref/src/menus.rs
@@ -1145,7 +1145,7 @@ fn settings_rows(
             )
             .changed()
         {
-            speed.multiplier = state.sim_speed;
+            speed.multiplier = state.sim_speed as f32;
         }
     } else {
         ui.add(
